# backend/openapi.internal.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SaaS Boilerplate",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/organizations": {
      "get": {
        "tags": [
          "Organizations"
        ],
        "summary": "Get All Organizations",
        "operationId": "get_all_organizations_v1_organizations_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MyOrganizationOut"
                  },
                  "type": "array",
                  "title": "Response Get All Organizations V1 Organizations Get"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Organizations"
        ],
        "summary": "Create An Organization",
        "operationId": "create_an_organization_v1_organizations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationBase"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/v1/organizations/{identifier}": {
      "patch": {
        "tags": [
          "Organizations"
        ],
        "summary": "Patch An Organization",
        "operationId": "patch_an_organization_v1_organizations__identifier__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Identifier"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Organizations"
        ],
        "summary": "Retrieve An Organization",
        "operationId": "retrieve_an_organization_v1_organizations__identifier__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Organizations"
        ],
        "summary": "Delete An Organization",
        "operationId": "delete_an_organization_v1_organizations__identifier__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/organizations/{organization_id}/transfer-ownership": {
      "post": {
        "tags": [
          "Organizations"
        ],
        "summary": "Transfer Organization Ownership",
        "operationId": "transfer_organization_ownership_v1_organizations__organization_id__transfer_ownership_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Organization Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferOwnershipIn"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/organizations/{organization_id}/users": {
      "get": {
        "tags": [
          "Organizations"
        ],
        "summary": "Get Organization Users",
        "operationId": "get_organization_users_v1_organizations__organization_id__users_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Organization Id"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 5,
              "default": 10,
              "title": "Page Size"
            }
          },
          {
            "name": "page_number",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page Number"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Comma-separated list of fields to sort by.\n\nUse a leading `-` before a field name to indicate descending sort order.",
              "default": "id",
              "title": "Sort"
            },
            "description": "Comma-separated list of fields to sort by.\n\nUse a leading `-` before a field name to indicate descending sort order."
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search term. Case-insensitive match across searchable fields.",
              "title": "Q"
            },
            "description": "Search term. Case-insensitive match across searchable fields."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_UserOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/invitations": {
      "get": {
        "tags": [
          "Invitations"
        ],
        "summary": "List Invitations",
        "operationId": "list_invitations_v1_invitations_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InvitationOut"
                  },
                  "type": "array",
                  "title": "Response List Invitations V1 Invitations Get"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/v1/invitations/{invitation_id}": {
      "delete": {
        "tags": [
          "Invitations"
        ],
        "summary": "Revoke Invitation",
        "description": "Deletes the invitation; its link stops working immediately.",
        "operationId": "revoke_invitation_v1_invitations__invitation_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "invitation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Invitation Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "InvitationInviterOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "email"
        ],
        "title": "InvitationInviterOut"
      },
      "InvitationOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "role_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Role Ids"
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/InvitationRoleOut"
            },
            "type": "array",
            "title": "Roles"
          },
          "invited_by": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InvitationInviterOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "is_expired": {
            "type": "boolean",
            "title": "Is Expired",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "role_ids",
          "roles",
          "invited_by",
          "created_at",
          "expires_at",
          "is_expired"
        ],
        "title": "InvitationOut"
      },
      "InvitationRoleOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "InvitationRoleOut"
      },
...
      "MyOrganizationOut": {
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "is_owner": {
            "type": "boolean",
            "title": "Is Owner",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "created_at",
          "updated_at",
          "name",
          "id",
          "is_owner"
        ],
        "title": "MyOrganizationOut"
      },
      "NameDescriptionOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description"
        ],
        "title": "NameDescriptionOut"
      },
...
      "OrganizationBase": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "OrganizationBase"
      },
      "OrganizationOut": {
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "created_at",
          "updated_at",
          "name",
          "id"
        ],
        "title": "OrganizationOut"
      },
      "OrganizationPatch": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "type": "object",
        "title": "OrganizationPatch"
      },
...
      "PaginatedResponse_UserOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/UserOut"
            },
            "type": "array",
            "title": "Items"
          },
          "page_number": {
            "type": "integer",
            "title": "Page Number"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page_number",
          "page_size",
          "total"
        ],
        "title": "PaginatedResponse[UserOut]"
      },
...
      "RoleOut": {
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "organization_id": {
            "type": "integer",
            "title": "Organization Id"
          },
          "is_protected": {
            "type": "boolean",
            "title": "Is Protected",
            "default": false
          },
          "permissions": {
            "items": {
              "$ref": "#/components/schemas/NameDescriptionOut"
            },
            "type": "array",
            "title": "Permissions"
          }
        },
        "type": "object",
        "required": [
          "created_at",
          "updated_at",
          "name",
          "description",
          "id",
          "organization_id",
          "is_protected",
          "permissions"
        ],
        "title": "RoleOut"
      },
...
      "TransferOwnershipIn": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "TransferOwnershipIn"
      },
...
      "UserOut": {
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "locale": {
            "type": "string",
            "title": "Locale",
            "default": "da"
          },
          "theme": {
            "type": "string",
            "title": "Theme",
            "default": "system"
          },
          "terms_accepted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terms Accepted At"
          },
          "mfa_enabled": {
            "type": "boolean",
            "title": "Mfa Enabled",
            "default": false
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/RoleOut"
            },
            "type": "array",
            "title": "Roles"
          }
        },
        "type": "object",
        "required": [
          "created_at",
          "updated_at",
          "name",
          "email",
          "id",
          "locale",
          "theme",
          "terms_accepted_at",
          "mfa_enabled",
          "roles"
        ],
        "title": "UserOut"
      },
...
      "ErrorResponse": {
        "properties": {
          "time": {
            "format": "date-time",
            "title": "Time",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "method": {
            "title": "Method",
            "type": "string"
          },
          "error_code": {
            "title": "Error Code",
            "type": "string"
          },
          "msg": {
            "title": "Msg",
            "type": "string"
          }
        },
        "required": [
          "time",
          "path",
          "method",
          "error_code",
          "msg"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "ValidationDetail": {
        "properties": {
          "error_code": {
            "title": "Error Code",
            "type": "string"
          },
          "field": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Field",
            "type": "array"
          },
          "msg": {
            "title": "Msg",
            "type": "string"
          },
          "ctx": {
            "additionalProperties": true,
            "title": "Ctx",
            "type": "object"
          }
        },
        "required": [
          "error_code",
          "field",
          "msg",
          "ctx"
        ],
        "title": "ValidationDetail",
        "type": "object"
      },
      "ValidationErrorResponse": {
        "$defs": {
          "ValidationDetail": {
            "properties": {
              "error_code": {
                "title": "Error Code",
                "type": "string"
              },
              "field": {
                "items": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "integer"
                    }
                  ]
                },
                "title": "Field",
                "type": "array"
              },
              "msg": {
                "title": "Msg",
                "type": "string"
              },
              "ctx": {
                "additionalProperties": true,
                "title": "Ctx",
                "type": "object"
              }
            },
            "required": [
              "error_code",
              "field",
              "msg",
              "ctx"
            ],
            "title": "ValidationDetail",
            "type": "object"
          }
        },
        "properties": {
          "time": {
            "format": "date-time",
            "title": "Time",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "method": {
            "title": "Method",
            "type": "string"
          },
          "error_code": {
            "title": "Error Code",
            "type": "string"
          },
          "msg": {
            "title": "Msg",
            "type": "string"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ValidationDetail"
            },
            "title": "Errors",
            "type": "array"
          }
        },
        "required": [
          "time",
          "path",
          "method",
          "error_code",
          "msg",
          "errors"
        ],
        "title": "ValidationErrorResponse",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "v1/auth/token"
          }
        }
      }
    }
  }
}